+++ python_modules/dagster/dagster/core/host_representation/__init__.py
@@ -10,7 +10,6 @@ from .external import (
     ExternalExecutionPlan,
     ExternalPartitionSet,
     ExternalPipeline,
-    ExternalPipelineExecutionResult,
     ExternalRepository,
     ExternalSchedule,
 )