+++ python_modules/dagster/dagster/core/host_representation/external.py
@@ -1,7 +1,6 @@
-from collections import OrderedDict, namedtuple
+from collections import OrderedDict
 
 from dagster import check
-from dagster.core.events import DagsterEvent
 from dagster.core.snap import ExecutionPlanSnapshot
 from dagster.core.utils import toposort
 
@@ -399,18 +398,6 @@ class ExternalPartitionSet:
     @property
     def pipeline_name(self):
         return self._external_partition_set_data.pipeline_name
-
-
-class ExternalPipelineExecutionResult(namedtuple("_ExternalPipelineExecutionResult", "event_list")):
-    def __new__(cls, event_list):
-        return super(ExternalPipelineExecutionResult, cls).__new__(
-            cls, check.list_param(event_list, "event_list", of_type=DagsterEvent)
-        )
-
-    @property
-    def success(self):
-        """bool: Whether all steps in the execution were successful."""
-        return all([not event.is_failure for event in self.event_list])
 
 
 class ExternalExecutable: